+++ presto-main/src/main/java/com/facebook/presto/connector/system/TaskSystemTable.java
@@ -142,7 +142,7 @@ public class TaskSystemTable
         if (duration == null) {
             return null;
         }
-        return (long) duration.toMillis();
+        return duration.toMillis();
     }
 
     private Long toBytes(DataSize dataSize)
