+++ presto-main/src/main/java/com/facebook/presto/metadata/ShardManagerDao.java
@@ -229,7 +229,7 @@ public interface ShardManagerDao
                 }
                 catch (UnableToObtainConnectionException e) {
                     log.warn("Failed to connect to database. Will retry again in %s. Exception: %s", delay, e.getMessage());
-                    Thread.sleep((long) delay.toMillis());
+                    Thread.sleep(delay.toMillis());
                 }
             }
         }
