+++ presto-main/src/main/java/com/facebook/presto/storage/StorageDao.java
@@ -55,7 +55,7 @@ public interface StorageDao
                 }
                 catch (UnableToObtainConnectionException e) {
                     log.warn("Failed to connect to database. Will retry again in %s. Exception: %s", delay, e.getMessage());
-                    Thread.sleep((long) delay.toMillis());
+                    Thread.sleep(delay.toMillis());
                 }
             }
         }
